# im-backoffice/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.Response:
    properties:
      key:
        type: string
      message:
        type: string
    type: object
...
  logisticCenter.LGLocationInfoDTO:
    properties:
      id:
        type: string
      location_name:
        type: string
    type: object
  logisticCenter.LGWarehouseInfoDTO:
    properties:
      id:
        type: string
      locations:
        items:
          $ref: '#/definitions/logisticCenter.LGLocationInfoDTO'
        type: array
      warehouse_name:
        type: string
    type: object
  logisticCenter.LogisticCenterDTO:
    properties:
      logistic_center_id:
        type: string
      logistic_center_name:
        type: string
    required:
    - logistic_center_name
    type: object
  logisticCenter.LogisticCenterRequest:
    properties:
      logistic_center_name:
        type: string
    required:
    - logistic_center_name
    type: object
...
  logisticCenter.LogisticCenterResponseDTO:
    properties:
      id:
        type: string
      logistic_center_name:
        type: string
      warehouses:
        items:
          $ref: '#/definitions/logisticCenter.LGWarehouseInfoDTO'
        type: array
    type: object
  user.RolesResp:
    properties:
      id:
        type: string
      role_type:
        type: string
    type: object
  user.UpdateUserReq:
    properties:
      enabled:
        type: boolean
      firstname:
        type: string
      lastname:
        type: string
      role_id:
        type: string
    type: object
  user.UserDTO:
    properties:
      enabled:
        type: boolean
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      role_id:
        type: string
      username:
        type: string
    type: object
  user.UserMeResp:
    properties:
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      role:
        $ref: '#/definitions/user.RolesResp'
      username:
        type: string
    type: object
  user.UserResp:
    properties:
      enabled:
        type: boolean
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      role:
        $ref: '#/definitions/user.RolesResp'
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an inventory management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inventory Management API
  version: "2.0"
paths:
...
  /logistic-center:
    post:
      consumes:
      - application/json
      parameters:
      - description: LogisticCenter
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/logisticCenter.LogisticCenterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/logisticCenter.LogisticCenterDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new logisticCenter.
      tags:
      - logistic-center
  /logistic-center/{logisticCenterId}:
    delete:
      parameters:
      - description: LogisticCenter ID
        in: path
        name: logisticCenterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new logisticCenter.
      tags:
      - logistic-center
    get:
      consumes:
      - application/json
      parameters:
      - description: LogisticCenter ID
        in: path
        name: logisticCenterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/logisticCenter.LogisticCenterResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Gets an logisticCenter by ID.
      tags:
      - logistic-center
    put:
      consumes:
      - application/json
      parameters:
      - description: LogisticCenter DTO
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/logisticCenter.LogisticCenterRequest'
      - description: LogisticCenter ID
        in: path
        name: logisticCenterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/logisticCenter.LogisticCenterResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Update the logisticCenter.
      tags:
      - logistic-center
  /logistic-centers:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logisticCenter.LogisticCenterResponseDTO'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all logisticCenters.
      tags:
      - logistic-center
  /logistic-centers/{page}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/logisticCenter.LogisticCenterResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all logisticCenters with pagination.
      tags:
      - logistic-center
  /user:
    post:
      consumes:
      - application/json
      description: Create a user in database
      parameters:
      - description: UserDTO
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/user.UserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a user.
      tags:
      - user
  /user/{id}:
    put:
      consumes:
      - application/json
      description: Update a user in database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateUserReq
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResp'
        "304":
          description: Not Modified
          schema:
            $ref: '#/definitions/errors.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - user
  /user/me:
    get:
      consumes:
      - '*/*'
      description: get user info from database by token in the request header
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserMeResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user info by token in the request header.
      tags:
      - user
  /user/roles:
    get:
      consumes:
      - '*/*'
      description: get all roles from roles table in database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.RolesResp'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all roles.
      tags:
      - user
  /user/roles/{id}:
    get:
      consumes:
      - '*/*'
      description: get one role by id from database
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.RolesResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get one role.
      tags:
      - user
  /users:
    get:
      consumes:
      - '*/*'
      description: get all users from users table in database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserResp'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all users.
      tags:
      - user
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"